waitUntilExit: ignore RunLoop.run()'s return value
We've been seeing frequent (but not reproducible in isolation) failures
in our builds where waitUntilExit returns but terminationStatus fails
its precondition on hasFinished because isRunning is still true.
This is a speculative fix, since I haven't been able to reproduce the
failure in a self-contained test.

This should be safe since Process.waitUntilExit isn't spec'd to relate
to RunLoop. In particular there is no guarantee that "stopping" a
RunLoop should cause waitUntilExit to return before the monitored
process has exited.