Change equalityOracle to compare based on values rather than indices
It appears that the `checkHashable` is mistakenly using an `equalityOracle` that merely compares the two index arguments themselves, rather than the underlying values at those index arguments. And so the comparison will always pass, regardless of whether the underlying values were different.

This appears to be a problem throughout the test cases, but this PR only fixes the check in TestDate.