Foundation: correct directory iteration on Windows
The path that was being constructed would elide the penultimate arc in
the path as the base URL was not marked as a directory.  As such, it was
assumed to be a file URL, and making a URL relative to it would truncate
the previously last arc.  Append the path component instead and
explicitly indicate if it is a directory component when building the URL
as we already have the information on hand.

This repairs the directory traversal on Windows.  The bug was identified
by the DocC test suite on Windows.