Foundation: support file system representation for long paths
This adjusts the file system representation accessor for a URL to
support paths beyond `MAX_PATH`.  We do not know what the use of the
file system representation is, so we cannot provide the NT style path
(`\\?\`-prefixed).  Rather than using the filled in buffer, resolve the
URL and then build the complete path copying that into a newly allocated
buffer.  This ensures that we always have sufficient space for the path
without having to grow the buffer.

Linux gets away with the current implementation by setting the
`PATH_MAX` to page size (4096).  There may be some value in using the
single implementation across the different platforms, but this will
allow the Windows path to make progress even if it diverges from the
reference implementation.

The issue was identified by the DocC test suite which quickly exceeds
the length.