Foundation: various improvements to long path support
Refactor directory iteration to avoid replicating the conversion to the
NT style paths everywhere.  The iteration logic can be done at a single
site with a callback to handle the iteration.

Improve handle long file paths on Windows when performing a `removeItem`
call.  Furthermore, correct a few areas where we were mishandling
junctions.  This would result in an early termination of the loop in
`removeItem` causing us to fail to clean up directories which we should
have been able to.  This improves the DocC test coverage on Windows.

Rework the file attribute reading on Windows operation to reformulate
the path to the absolute path representation and then into the NT form
before performing the operation.  This technically is a partial repair
as drive relative paths where the current directory is deep enough, the
path evaluation would fail.

Co-authored-by: Gwynne Raskind <gwynne@darkrainfall.org>