build: adjust the build for rebranch (#4820)
The newer clang compiler prefers `module.modulemap` over `module.map`.
This will result in us using the framework form of the module
definitions for CoreFoundation. While this is fine for the compilation,
the linker directive to link against the framework will also be emitted
into the object files. This causes a linker failure on Windows as
`-framework (CoreFoundation|CFURLInterface|CFXMLInterface)` is not a
valid linker option. Use the swift control over the directives to avoid
emitting the framework link directive to repair the build.