[wasm] Port thread-related APIs for no thread platform
This change ports the thread-related APIs for the no thread platform by
adding no-op implementations. Note that WASI is going to support threads
in the near future, so this does not guard by `TARGET_OS_WASI` but by
`!SWIFT_CORELIBS_FOUNDATION_HAS_THREADS` to make it easier for future
WASI + threads support.