Fixes issue #4642: Encountering "Constant strings cannot be deallocated" in Linux foundation (#4653)
The reported error was an CFSTR("") which is later released. While this should
not be a problem, the Linux implementation of CFSTR does not ignore dealloc
on constant strings. Fixed this by calling CFRetain on the constant string.

Strictly speaking this is only a workaround. Issue #1351 has some hints how
this could be fixed but the workaround is used a over the code so I think it
is okay to use it here, too.

I found the same problem in CFDateIntervalFormatter.c where it appeared in a
error handling code path that should never be called. Fixed anyways.