Foundation: avoid temporary file leaks on Windows
Correct the implementation of `_NSCleanupTemporaryFile` to use
`MoveFileExW` with `MOVEFILE_COPY_ALLOWED` to prefer to use file renames
rather than copy and delete.  This should improve performance in the
general single volume case and still allow the correct behaviour in the
multi-volume scenarios using `CopyFile` and `DeleteFile`.

Thanks to @darinf for identifying the test case with 65535 temporary
files!