[wasm] Port BlocksRuntime for no dlfcn.h platforms
BlocksRuntime uses dlsym to find objc_destructInstance, which is not
available on all platforms. This change adds a check for dlfcn.h and
only uses dlsym if it is available and otherwise crashes the program.

The crash will not usually happen, as `_Block_use_RR` is only called
from objc-auto, which is not available for such platforms.