[wasm] Import `sys/mman.h` through CoreFoundation
Importing `wasi_emulated_mman` in Swift code requires the client to
define `_WASI_EMULATED_MMAN`. `@_implementationOnly import` was the only
way to work around this, but it's now declared as unsafe if it's used in
non-resilient moules (even though it's actually safe as long as the
module does not use type layout of the imported module).
The new scoped import feature always requires the transitive clients to
load privately imported modules, so it's not a good fit for this tricky
case where a special macro must be defined before importing the module.

This patch imports `sys/mman.h` through CoreFoundation while defining
`_WASI_EMULATED_MMAN` in the header exposed by modulemap. This way, the
clients don't need to define it.