Removing checks for deleted timezones
Recent Linux distributions no longer carry some deprecated timezones.
Removing them, or replacing with the equivalent timezone, according to
the contents of `/usr/share/zoneinfo` across the years of Linux that we
are currently supporting.

Removed timezones include:
 - Israel          -> Asia/Jerusalem
 - US/Pacific      -> America/Los_Angeles
 - Brazil/East     -> America/Sao_Paulo

Unfortunately, the 'Europe/Kyiv'/'Europe/Kiev' timezone has changed
around a bit in the past few years, likely as a result of political
events. As a result, I don't think I have a single spelling that will
work across all of the versions of OS's that we are trying to support.
If we really want to ensure that the tests are testing that Foundation
can load timezone information and not test the current political events,
it would likely be a good idea to save a couple of timezone files off
for testing purposes and use those across OS versions instead of hoping
that the list of timezones doesn't change from year-to-year. For now,
I'm going to point the test at London instead. The daylight savings time
shifted on the same date in 2020, so there is no other change to the
test needed.