CMake: Fix missing librt target
When building the static SDK, dispatch is built as a static archive, so
it's dependency on librt is exposed as part of its link interface.
This shows up in the `dispatchConfig.cmake` file imported by
`find_package(dispatch CONFIG)`.

```
CMake Error at build/x86_64/dispatch/cmake/modules/dispatchExports.cmake:71 (set_target_properties):
  The link interface of target "dispatch" contains:

    RT::rt

  but the target was not found.  Possible reasons include:

    * There is a typo in the target name.
    * A find_package call is missing for an IMPORTED target.
    * An ALIAS target is missing.
```

It would be great if the dispatch config could include that target to
ensure that everyone is using the same librt, it does not, so we have to
find and include it ourselves.