Bring back defines and flags
Several definitions and flags were dropped from the build when moving to
swift-froundation.
This patch puts back the ones that we need in order to build against
musl.

Since the musl builds are static, we also need to pick up the link to
libRT or dispatch will fail to link.