[6.0.2] Add a thread-safe implementation of Process.currentDirectoryURL (#5095)
* Add a thread-safe implementation of Process.currentDirectoryURL

posix_spawn_file_actions_addchdir_np is the official way to set the working directory of a spawned process and is supported on both macOS and glibc (Linux, etc.). This makes Process.currentDirectoryURL thread-safe, as the current approach will result in the working directory being nondeterministically assigned when spawning processes across multiple threads, using different working directories.

* Restore thread-unsafe fallback for setting the Process working directory

Swift still needs to support Amazon Linux 2 until it EoLs in mid-2025. So restore the thread-unsafe fallback for systems with glibc older than version 2.29, which was removed in #4981.