Ensure strsep operates on a copy of PATH. (#5289)
strsep modifies its first argument, which means that if we don't make a
copy, any PATH processing that may occur after _CFProcessPath executes
breaks because strsep will effectively change PATH to be only the first
list item. For example, swiftpm's tool finding fails because it searches
PATH, and if the tool in question doesn't reside in the first PATH item,
things break.

This only occurs on BSD and only when the program name, i.e., argv[0], is
not an absolute path.