Fix socket lifecycle management
Motivation:

When an `easy_handle` is added to a `multi_handle`, the callback set for `CURLOPT_CLOSESOCKETFUNCTION` constructs a `_MultiHandle` from an unretained reference to `self`. This could cause a crash in `URLSession` if the `_MultiHandle`'s `self` is deinitialised before the callback is called, and the socket will never be closed.

_Partially resolves #3813_

Modifications:

- Set the `CURLOPT_CLOSESOCKETFUNCTION` callback to do nothing, deferring closing the socket.
- Schedule a socket for closure in the cancellation handler of its dispatch sources.

Result:

Better management of socket lifecycle.

Testing:

Passing existing tests is sufficient.